// src/payload-types.ts
export interface Post {
  id: number;
  title: string;
  /**
   * Used in the URL, e.g. /posts/my-post
   */
  slug: string;
  categories?: (number | Category)[] | null;
  date: string;
  author?: string | null;
  /**
   * Short summary shown on listing pages.
   */
  excerpt?: string | null;
  content?: string | null;
  featuredImage?: (number | null) | Media;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  title: string;
  /**
   * Used in URLs, e.g. /categories/power-platform
   */
  slug: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}